Style and organize calculator keys
I used flexbox to arrange the sets of keys contained in their respective
divs to make the calculator look like a calculator. Using widths, heights,
and a few other flexbox tricks I now have a calculator layout. I also
used CSS to change the colors of the keys based on their respective
category. This will change in future edits to give it cohesion.