Add functionality to calculator
I used a switch function to set up my calculator button functions based on
button attributes. This was to make sure that a user can enter a number,
an operator, and another number, then hit = and the operation function
should fire based on this information and correctly produce an answer.

The answer is also displayed now in one of two divs. The calculation will
be on the bottom div and when they hit = the answer will be above the
equation.

I could not figure out how to do negative numbers.